init markers
