+++ src/routes/+layout.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+	import '../global.css';
+	let { children } = $props();
+</script>
+
+{@render children()}
